Complete rework. Really need to check-in more often :/
Added phone version, with its own phonePageManager.
Refactored all the xml to properly use Android's size buckets.
Changed the loads to all be on demand.
All fragments get loaded up at start and are shown and hidden, not add/removed.
The booth flip is moved into a static helper so it can be used by phone and tablet.
Optimized xml UI from LinearLayouts with weights to RelativeLayout (it measures twice for each child, and so grows exponentially in measure time)