Changed the app to show the 'Phone' version to Large and below (was looking crap on 7" Nexus)
Changed the app to have a shipped database which is copied on first load.
Removed sync capability for now (commented out in RealcommApplication) while about to start work on writing custom beacon service. (Radius Networks' does too much on the UIThread when there are lots of beacons and doesn't allow the sync to function properly)