Whew, epic check-in. Changed name from Realcomm3 to Realcomm, which means just about every file was touched.
Wish it wasn't like this, but there are a number of bug fixes included in here, time was pressing, and I wasn't sure enough about the name change to check-in yet.
Fixes include (but not limited to):
Changed from timer to handler on BoothFlipper
Crash from fragment doing commit() after onPause (open Google user from settings, turn bluetooth off)
Minor tweaks and changes to the look and feel, padding here, margin there, layout for 7" tabs in the middle.