Linkify was breaking contacts (open email and come back would then retain last text). Split up contact details and created custom intents for each one.
Changed LineIndicator to CircleIndicator
Added .commitAllowingStateLoss() everywhere, as it was causing crashes. Not the best solution, but will do for now.
Updated the schedule fragment to start showing the schedule for the current day.
Updated the booth fragments ellipsizing. No longer using ViewTreeObserver, didn't feel right.
InfoFragment LastUpdated is now updated when the download is finished.